requirements.txt
# AXCSAS Dependencies

# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Plotting
matplotlib>=3.4.0

# Configuration
pyyaml>=5.4.0

# Testing
pytest>=6.2.0

# Optional: Jupyter Notebooks
jupyter>=1.0.0
